Finished claim-pending-rewards. Fixed rewards.
