+++ clarity/contracts/dao.clar
@@ -232,4 +232,18 @@
       active: true
     }
   )
+
+  (map-set contracts
+    { name: "vault-data" }
+    {
+      address: 'STSTW15D618BSZQB85R058DS46THH86YQQY6XCB7,
+      qualified-name: 'STSTW15D618BSZQB85R058DS46THH86YQQY6XCB7.vault-data
+    }
+  )
+  (map-set contracts-data
+    { qualified-name: 'STSTW15D618BSZQB85R058DS46THH86YQQY6XCB7.vault-data }
+    {
+      active: true
+    }
+  )
 )
