+++ clarity/tests/freddie_test.ts
@@ -313,60 +313,3 @@ Clarinet.test({
     block.receipts[0].result.expectErr().expectUint(98); // wrong token error
   }
 });
-
-Clarinet.test({
-  name: "freddie: rewards",
-  async fn(chain: Chain, accounts: Map<string, Account>) {
-    let deployer = accounts.get("deployer")!;
-    let wallet_1 = accounts.get("wallet_1")!;
-    let block = chain.mineBlock([
-      
-      Tx.contractCall("oracle", "update-price", [
-        types.ascii("STX"),
-        types.uint(77),
-      ], deployer.address),
-      Tx.contractCall("freddie", "collateralize-and-mint", [
-        types.uint(5000000),
-        types.uint(1925000),
-        types.principal(deployer.address),
-        types.ascii("STX-A"),
-        types.ascii("STX"),
-        types.principal("STSTW15D618BSZQB85R058DS46THH86YQQY6XCB7.stx-reserve"),
-        types.principal(
-          "STSTW15D618BSZQB85R058DS46THH86YQQY6XCB7.arkadiko-token",
-        ),
-      ], deployer.address),
-      Tx.contractCall("freddie", "collateralize-and-mint", [
-        types.uint(5000000),
-        types.uint(1925000),
-        types.principal(deployer.address),
-        types.ascii("STX-A"),
-        types.ascii("STX"),
-        types.principal("STSTW15D618BSZQB85R058DS46THH86YQQY6XCB7.stx-reserve"),
-        types.principal(
-          "STSTW15D618BSZQB85R058DS46THH86YQQY6XCB7.arkadiko-token",
-        ),
-      ], deployer.address),
-    ]);
-
-    let call = chain.callReadOnlyFn("vault-data", "get-pending-rewards", [types.principal(deployer.address)], deployer.address);
-    call.result.expectOk().expectUint(1)
-
-    call = chain.callReadOnlyFn("vault-data", "get-collateral-of-vault-by-id", [types.uint(1)], deployer.address);
-    call.result.expectUint(5000000)
-
-    call = chain.callReadOnlyFn("vault-data", "get-total-collateral-of", [types.principal(deployer.address)], deployer.address);
-    call.result.expectUint(10000000)
-    
-    chain.mineEmptyBlock(1);
-
-    call = chain.callReadOnlyFn("vault-data", "get-pending-rewards", [types.principal(deployer.address)], deployer.address);
-    call.result.expectOk().expectUint(2)
-
-    chain.mineEmptyBlock(12*30*144);
-
-    call = chain.callReadOnlyFn("vault-data", "get-pending-rewards", [types.principal(deployer.address)], deployer.address);
-    call.result.expectOk().expectUint(5184200)
-    
-  },
-});
