+++ clarity/tests/vault-data_test.ts
@@ -1,0 +1,107 @@
+import {
+  Account,
+  Chain,
+  Clarinet,
+  Tx,
+  types,
+} from "https://deno.land/x/clarinet@v0.6.0/index.ts";
+
+
+Clarinet.test({
+  name: "vault-data: vault DIKO rewards",
+  async fn(chain: Chain, accounts: Map<string, Account>) {
+    let deployer = accounts.get("deployer")!;
+    let wallet_1 = accounts.get("wallet_1")!;
+    let block = chain.mineBlock([
+      
+      Tx.contractCall("oracle", "update-price", [
+        types.ascii("STX"),
+        types.uint(77),
+      ], deployer.address),
+      Tx.contractCall("freddie", "collateralize-and-mint", [
+        types.uint(5000000),
+        types.uint(1925000),
+        types.principal(deployer.address),
+        types.ascii("STX-A"),
+        types.ascii("STX"),
+        types.principal("STSTW15D618BSZQB85R058DS46THH86YQQY6XCB7.stx-reserve"),
+        types.principal(
+          "STSTW15D618BSZQB85R058DS46THH86YQQY6XCB7.arkadiko-token",
+        ),
+      ], deployer.address),
+      Tx.contractCall("freddie", "collateralize-and-mint", [
+        types.uint(5000000),
+        types.uint(1925000),
+        types.principal(deployer.address),
+        types.ascii("STX-A"),
+        types.ascii("STX"),
+        types.principal("STSTW15D618BSZQB85R058DS46THH86YQQY6XCB7.stx-reserve"),
+        types.principal(
+          "STSTW15D618BSZQB85R058DS46THH86YQQY6XCB7.arkadiko-token",
+        ),
+      ], deployer.address),
+    ]);
+
+    let call = chain.callReadOnlyFn("vault-data", "get-collateral-of-vault-by-id", [types.uint(1)], deployer.address);
+    call.result.expectUint(5000000)
+
+    call = chain.callReadOnlyFn("vault-data", "get-total-collateral-of", [types.principal(deployer.address)], deployer.address);
+    call.result.expectUint(10000000)
+
+    // Check rewards
+    call = chain.callReadOnlyFn("vault-data", "get-pending-rewards", [types.principal(deployer.address)], deployer.address);
+    call.result.expectOk().expectUint(55123110)
+    
+    chain.mineEmptyBlock(1);
+
+    call = chain.callReadOnlyFn("vault-data", "get-pending-rewards", [types.principal(deployer.address)], deployer.address);
+    call.result.expectOk().expectUint(110246230)
+
+    chain.mineEmptyBlock(12*30*144);
+
+    call = chain.callReadOnlyFn("vault-data", "get-pending-rewards", [types.principal(deployer.address)], deployer.address);
+    call.result.expectOk().expectUint(1483801920350)
+    
+  },
+});
+
+Clarinet.test({
+  name: "vault-data: claim DIKO rewards",
+  async fn(chain: Chain, accounts: Map<string, Account>) {
+    let deployer = accounts.get("deployer")!;
+    let wallet_1 = accounts.get("wallet_1")!;
+    let block = chain.mineBlock([
+      
+      Tx.contractCall("oracle", "update-price", [
+        types.ascii("STX"),
+        types.uint(77),
+      ], deployer.address),
+      Tx.contractCall("freddie", "collateralize-and-mint", [
+        types.uint(5000000),
+        types.uint(1925000),
+        types.principal(deployer.address),
+        types.ascii("STX-A"),
+        types.ascii("STX"),
+        types.principal("STSTW15D618BSZQB85R058DS46THH86YQQY6XCB7.stx-reserve"),
+        types.principal(
+          "STSTW15D618BSZQB85R058DS46THH86YQQY6XCB7.arkadiko-token",
+        ),
+      ], deployer.address),
+    ]);
+
+    chain.mineEmptyBlock(30);
+
+    let call = chain.callReadOnlyFn("vault-data", "get-pending-rewards", [types.principal(deployer.address)], deployer.address);
+    call.result.expectOk().expectUint(1708816595)
+
+    call = chain.callReadOnlyFn("arkadiko-token", "get-balance-of", [types.principal(wallet_1.address)], wallet_1.address);
+    call.result.expectOk().expectUint(150000000000);   
+
+    call = chain.callReadOnlyFn("vault-data", "claim-pending-rewards", [], deployer.address);
+    call.result.expectOk().expectUint(1708816595)
+
+    call = chain.callReadOnlyFn("arkadiko-token", "get-balance-of", [types.principal(wallet_1.address)], wallet_1.address);
+    call.result.expectOk().expectUint(150000000000);  
+
+  },
+});
